Rework the clipvalues passed directly to the template to all the views, not only edit.
IT'S RESPONSIBILITY OF THE DEVELOPER TO USE THEM SAFELY.
GET parameters or arguments with name `_namehere` will be passed to the template
and will be available at `$clipvalues.namehere`.

Arguments works like `{clip_func func='list' _param1='value1' ___param2='Translatable Value 2' ...}`
and will be available in the template with `$clipvalues.param1` and `$clipvalues.param2`.