Check the templates existance before the database processing.
Also enables a general template in the templates/ folder for list and display,
to be used widely by all the pubtypes if it exists.

If /pubtype/list_myoutput.tpl does not exist but /list_myoutput.tpl does
the last general one will be used. Shared for all the pubtypes if the requested one does not exists.